backend/requirements.txt
# =============================================================================
# Virtú Backend - Requirements
# Use Python 3.11 ou 3.12
# =============================================================================

Django>=5.0,<6.0
wagtail>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
whitenoise>=6.6,<7.0
requests>=2.31,<3.0
